WebContext: add `getContent` for a11y-like content tree
This should be usable in snapshot testing.

e.g. Instead of making lots of small calls to confirm that
certain labels, widgets, etc. are on the page the entire content
tree can be dumped to a JSON serializable tree and used as a
baseline to compare against, either the entire tree, or some
subset(s).

Note: This tree's schema may change over time as support for
ARIA attributes is not yet part of WebAutomator.