Serialize WebElements inside of a top-level array to appease iOS driver
`WebDriver.executeScript` will not work with iOS whenever a DOM
element is returned inside of an Object. DOM elements can be returned
individually (i.e. the return value is simply a single `WebElement`) and
can be returned as an item in an array but it cannot be returned in an
array of arrays or inside of an object.

This change replaces all elements from the content array with
"references" to those elements and then bundles the modified content
with the elements to be returned as a single array (so that DOM elements
serialize correctly). After deserialization these references are then
replaced with the actual `WebElement` instances to reconstruct the
original content array with elements inlined.

Note: This same issued exists with previous GeckoDriver and Firefox
combinations.